+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * **Breaking change**: Handling of AcroForm check boxes to allow multiple
   widgets with different values
+* CLI command `hexapdf form` to support new check box features
 * [HexaPDF::Content::Canvas#text] to use the font size as leading if no leading
   has been set
 * [HexaPDF::Content::Canvas#line_with_rounded_corner] to be a public method